FEAT: Second version, split functionality in files
