+++ include/erdsdk.h
@@ -7,11 +7,12 @@
 #include "transaction/messagesigner.h"
 #include "transaction/transaction.h"
 #include "transaction/esdt.h"
-#include "smartcontracts/sc_arguments.h"
-#include "smartcontracts/contract_call.h"
 #include "transaction/gas_estimator.h"
 #include "transaction/token_payment.h"
 #include "transaction/payload_builder.h"
+#include "transaction/transaction_factory.h"
+#include "smartcontracts/sc_arguments.h"
+#include "smartcontracts/contract_call.h"
 #include "account/account.h"
 #include "account/address.h"
 #include "filehandler/isecretkey.h"
