+++ include/transaction/transaction_builder_esdt.h
@@ -1,0 +1,31 @@
+#ifndef ERD_TRANSACTION_BUILDER_ESDT_H
+#define ERD_TRANSACTION_BUILDER_ESDT_H
+
+#include "transaction_builder_input.h"
+#include "token_payment.h"
+#include "itransaction_builder.h"
+#include "transaction/payload_builder.h"
+#include "transaction/transaction.h"
+
+class TransactionFactory;
+
+class TransactionESDTBuilder : public ITransactionBuilder
+{
+    friend class TransactionFactory;
+
+public:
+
+    TransactionESDTBuilder &withContractCall(ContractCall contractCall);
+
+    Transaction build() override;
+
+private:
+    explicit TransactionESDTBuilder(const TransactionBuilderInput& txInput, TokenPayment payment);
+
+    TransactionBuilderInput m_txInput;
+    TokenPayment m_tokenPayment;
+    ContractCall m_contractCall;
+};
+
+
+#endif //ERD_TRANSACTION_BUILDER_ESDT_H
