+++ src/transaction/transaction_builder_esdt.cpp
@@ -1,0 +1,39 @@
+#include <utility>
+
+#include "transaction/transaction_builder_esdt.h"
+
+
+TransactionESDTBuilder::TransactionESDTBuilder(const TransactionBuilderInput& txInput, TokenPayment payment) :
+        m_txInput(txInput), m_tokenPayment(std::move(payment)), m_contractCall("")
+{}
+
+TransactionESDTBuilder& TransactionESDTBuilder::withContractCall(ContractCall contractCall)
+{
+    m_contractCall = std::move(contractCall);
+    return *this;
+}
+
+Transaction TransactionESDTBuilder::build()
+{
+    std::string payload = ESDTTransferPayloadBuilder()
+            .setPayment(m_tokenPayment)
+            .withContractCall(m_contractCall)
+            .build();
+
+    uint64_t gasLimit = m_txInput.gasEstimator.forESDTTransfer(payload.size());
+
+    return Transaction(
+            m_txInput.nonce,
+            "0",
+            m_txInput.receiver,
+            m_txInput.sender,
+            nullptr,
+            nullptr,
+            m_txInput.gasPrice,
+            gasLimit,
+            std::make_shared<bytes>(bytes(payload.begin(), payload.end())),
+            nullptr,
+            m_txInput.chainID,
+            DEFAULT_VERSION,
+            DEFAULT_OPTIONS);
+}
